+++ src/packages/eslint-config-kiwicom/ourRules.js
@@ -640,4 +640,5 @@ module.exports = {
   // Kiwi.com custom rules
   'kiwicom-incubator/only-nullable-fields': ERROR,
   'kiwicom-incubator/no-invalid-flow-annotations': ERROR,
+  'kiwicom-incubator/no-internal-flow-type': NEXT_VERSION_ERROR,
 };